requirements.txt
asgiref==3.9.0
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-widget-tweaks==1.5.0
Faker==37.4.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
